No more System.Drawing
Vendor a snapshot of ImageSharp, and quick-build it into an assembly
for my use. If we find bugs we'll fix them inline and then contribute
the patch back up.

Also enhance the Honeycomb sink so that I can specify the target
dataset when I emit the record, and then re-do the thumbnail logging
to have a regular structure which will fit into the honeycomb model a
little better.

This will also make it easier to query and the like; I am happy with
it so far! It's also absolutely necessary, so that I can watch for
bugs in ImageSharp and have a hope of diagnosing why I didn't get a
thumbnail...

(As a side effect I can probably run on dotnet core now.)