More mobile work.
I'm really starting to want to go to a dedicated CSS option so that I
can style the landing page without hackery. I think I might have to do
that so I can build the landing, signup, and login pages to be
consistent with the rest of the app.

Or maybe I should just hack it through and release and fix it
later. Probably the latter.