Build my own skyliner
skyliner.io is shutting down. But that's OK, we can put on our own
show right here in the barn!

There are a million things wrong with this, but the big ones are:

- Haven't implemented build automation yet.

- The cloudformation template that I generate still has some QA
  specific stuff in it.

- This only does release stacks, no environment stacks or base
  stacks. I still don't fully understand that stuff, so let's just
  hope that the skyliner folks were pretty smart when they set it up.

The first one will be done before this whole project can really be
called done, the second one is just bugs that will be fixed when we
try to deploy production with it.

The third one-- I don't really need automation for that stuff right
now since I only have the one app and I'm not building for
multi-tenancy. I'll only fix that one if I decide I need to go into
another region.